Rename analyze folder
