+++ src/analyzer/analyze.test.ts
