+++ src/analyzer/constants.ts
