+++ src/analyzer/index.ts
