+++ src/analyzer/model.ts
