+++ src/analyzer/util.ts
