+++ src/components/Providers.tsx
@@ -13,7 +13,7 @@ import {
   OptionsContext,
 } from "../contexts/OptionsContext"
 import { SearchContext } from "../contexts/SearchContext"
-import { HydratedGitBlobObject, AnalyzerData } from "~/analyze/model"
+import { HydratedGitBlobObject, AnalyzerData } from "~/analyzer/model"
 import { MetricContext } from "../contexts/MetricContext"
 import { DataContext } from "../contexts/DataContext"
 import { addAuthorUnion, makeDupeMap } from "../authorUnionUtil"
