+++ src/components/Tooltip.tsx
@@ -2,7 +2,7 @@ import { Box, BoxSubTitle, LegendDot } from "./util"
 import { useMouse } from "react-use"
 import { useMemo, useRef } from "react"
 import styled from "styled-components"
-import { HydratedGitBlobObject } from "~/analyze/model"
+import { HydratedGitBlobObject } from "~/analyzer/model"
 import { useOptions } from "../contexts/OptionsContext"
 import { Spacer } from "./Spacer"
 import { useMetricCaches } from "../contexts/MetricContext"
